delete old dialog, rename dialog next to dialog
