--- packages/lab/src/dialog-next/DialogNext.css
@@ -1,63 +1,0 @@
-/* Styles applied to DialogNext component */
-.saltDialogNext {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  padding-top: var(--salt-spacing-300);
-  padding-bottom: var(--salt-spacing-300);
-  background: var(--salt-container-primary-background);
-  max-height: 100%;
-  max-width: 100%;
-  box-shadow: var(--salt-overlayable-shadow-modal);
-  min-width: calc(10 * var(--salt-size-base));
-  max-width: calc(20 * var(--salt-size-base));
-  z-index: calc(var(--salt-zIndex-appHeader) - 1);
-  overflow-y: auto;
-}
-
-/* Styles applied to DialogNext background overlay */
-.saltDialogNext-overlay {
-  background: var(--salt-overlayable-background);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: var(--salt-spacing-400);
-  z-index: var(--salt-zIndex-dialog);
-}
-
-.saltDialogNext-withStatus {
-  border-width: var(--salt-size-border);
-  border-style: var(--salt-container-borderStyle);
-}
-
-.saltDialogNext-info {
-  border-color: var(--salt-status-info-borderColor);
-}
-
-.saltDialogNext-error {
-  border-color: var(--salt-status-error-borderColor);
-}
-
-.saltDialogNext-warning {
-  border-color: var(--salt-status-warning-borderColor);
-}
-
-.saltDialogNext-success {
-  border-color: var(--salt-status-success-borderColor);
-}
-
-/* Styles applied when the component mounts */
-.saltDialogNext.saltDialogNext-enterAnimation {
-  animation: var(--salt-animation-fade-in-center);
-}
-
-/* Styles applied when the component unmounts */
-.saltDialogNext.saltDialogNext-exitAnimation {
-  animation: var(--salt-animation-fade-out-back);
-}
-
-.saltDialogNext .saltDialogNext-closeButton {
-  position: absolute;
-  top: 0;
-  right: 0;
-}
