--- packages/lab/src/dialog-next/DialogNextActions.css
@@ -1,7 +1,0 @@
-.saltDialogNextActions {
-  padding: var(--salt-spacing-300);
-  padding-bottom: 0;
-  display: flex;
-  justify-content: flex-end;
-  gap: var(--salt-spacing-100);
-}
