--- packages/lab/src/dialog-next/DialogNextTitle.css
@@ -1,38 +1,0 @@
-.saltDialogNextTitle {
-  display: flex;
-  gap: var(--salt-spacing-100);
-  padding-bottom: var(--salt-spacing-100);
-  padding-left: var(--salt-spacing-300);
-  padding-right: var(--salt-spacing-300);
-  margin: 0;
-}
-
-.saltDialogNextTitle-withAccent {
-  position: relative;
-}
-
-.saltDialogNextTitle-withAccent::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: var(--salt-spacing-100);
-  width: var(--salt-size-accent);
-  background: var(--salt-accent-background);
-}
-
-.saltDialogNextTitle-error::before {
-  background: var(--salt-status-error-borderColor);
-}
-
-.saltDialogNextTitle-info::before {
-  background: var(--salt-status-info-borderColor);
-}
-
-.saltDialogNextTitle-success::before {
-  background: var(--salt-status-success-borderColor);
-}
-
-.saltDialogNextTitle-warning::before {
-  background: var(--salt-status-warning-borderColor);
-}
