--- packages/lab/src/dialog-next/index.tsx
@@ -1,6 +1,0 @@
-export * from "./DialogNext";
-export * from "./DialogNextTitle";
-export * from "./DialogNextActions";
-export * from "./DialogNextContent";
-export * from "./DialogNextContext";
-export * from "./DialogNextCloseButton";
