+++ packages/lab/src/dialog/Dialog.css
@@ -1,28 +1,63 @@
+/* Styles applied to Dialog component */
 .saltDialog {
-  --dialog-background: var(--saltDialog-background, var(--salt-container-primary-background));
-  --dialog-borderBottomWidth: var(--saltDialog-borderBottomWidth, 2px);
-  --dialog-borderColor: var(--saltDialog-borderColor, var(--salt-container-primary-borderColor));
-  --dialog-borderStyle: var(--saltDialog-borderStyle, var(--salt-container-borderStyle));
-  --dialog-borderWidth: var(--saltDialog-borderColor, var(--salt-size-border));
-  --dialog-boxShadow: var(--saltDialog-boxShadow, var(--salt-overlayable-shadow-modal));
-  --dialog-margin: var(--saltDialog-margin, var(--salt-size-container-spacing));
-  --dialog-padding: var(--saltDialog-padding, var(--salt-size-container-spacing));
-}
-
-.saltDialog {
-  background: var(--dialog-background);
-  border-color: var(--dialog-borderColor);
-  border-width: var(--dialog-borderWidth);
-  border-style: var(--dialog-borderStyle);
-  border-bottom-width: var(--dialog-borderBottomWidth);
+  position: relative;
   display: flex;
   flex-direction: column;
-  inset: 0; /* what does this do ? */
+  padding-top: var(--salt-spacing-300);
+  padding-bottom: var(--salt-spacing-300);
+  background: var(--salt-container-primary-background);
   max-height: 100%;
   max-width: 100%;
-  overflow: auto;
-  padding: var(--dialog-padding);
-  position: relative;
-  box-shadow: var(--saltDialog-boxShadow, var(--dialog-boxShadow));
-  margin: var(--saltDialog-margin, var(--dialog-margin));
+  box-shadow: var(--salt-overlayable-shadow-modal);
+  min-width: calc(10 * var(--salt-size-base));
+  max-width: calc(20 * var(--salt-size-base));
+  z-index: calc(var(--salt-zIndex-appHeader) - 1);
+  overflow-y: auto;
+}
+
+/* Styles applied to Dialog background overlay */
+.saltDialog-overlay {
+  background: var(--salt-overlayable-background);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: var(--salt-spacing-400);
+  z-index: var(--salt-zIndex-dialog);
+}
+
+.saltDialog-withStatus {
+  border-width: var(--salt-size-border);
+  border-style: var(--salt-container-borderStyle);
+}
+
+.saltDialog-info {
+  border-color: var(--salt-status-info-borderColor);
+}
+
+.saltDialog-error {
+  border-color: var(--salt-status-error-borderColor);
+}
+
+.saltDialog-warning {
+  border-color: var(--salt-status-warning-borderColor);
+}
+
+.saltDialog-success {
+  border-color: var(--salt-status-success-borderColor);
+}
+
+/* Styles applied when the component mounts */
+.saltDialog.saltDialog-enterAnimation {
+  animation: var(--salt-animation-fade-in-center);
+}
+
+/* Styles applied when the component unmounts */
+.saltDialog.saltDialog-exitAnimation {
+  animation: var(--salt-animation-fade-out-back);
+}
+
+.saltDialog .saltDialog-closeButton {
+  position: absolute;
+  top: 0;
+  right: 0;
 }
