+++ packages/lab/src/dialog/DialogActions.css
@@ -1,0 +1,7 @@
+.saltDialogActions {
+  padding: var(--salt-spacing-300);
+  padding-bottom: 0;
+  display: flex;
+  justify-content: flex-end;
+  gap: var(--salt-spacing-100);
+}
