+++ packages/lab/src/dialog/DialogCloseButton.css
@@ -1,4 +1,4 @@
-.saltButton.saltDialogNextCloseButton {
+.saltButton.saltDialogCloseButton {
   position: absolute;
   top: 0;
   right: 0;
