+++ packages/lab/src/dialog/DialogCloseButton.tsx
@@ -5,18 +5,18 @@ import { useWindow } from "@salt-ds/window";
 import { useComponentCssInjection } from "@salt-ds/styles";
 import { CloseIcon } from "@salt-ds/icons";
 
-import { useDialogNextContext } from "./DialogNextContext";
-import dialogNextCloseButtonCss from "./DialogNextCloseButton.css";
+import { useDialogContext } from "./DialogContext";
+import dialogCloseButtonCss from "./DialogCloseButton.css";
 
-const withBaseName = makePrefixer("saltDialogNextCloseButton");
+const withBaseName = makePrefixer("saltDialogCloseButton");
 
-export const DialogNextCloseButton = forwardRef<HTMLButtonElement, ButtonProps>(
-  function DialogNextCloseButton({ className, ...rest }, ref) {
-    const { dialogId } = useDialogNextContext();
+export const DialogCloseButton = forwardRef<HTMLButtonElement, ButtonProps>(
+  function DialogCloseButton({ className, ...rest }, ref) {
+    const { dialogId } = useDialogContext();
     const targetWindow = useWindow();
     useComponentCssInjection({
-      testId: "salt-dialog-next-close-button",
-      css: dialogNextCloseButtonCss,
+      testId: "salt-dialog-close-button",
+      css: dialogCloseButtonCss,
       window: targetWindow,
     });
