--- packages/lab/src/dialog/index.ts
@@ -1,4 +1,0 @@
-export * from "./DialogActions";
-export * from "./DialogContent";
-export * from "./DialogTitle";
-export * from "./Dialog";
