+++ packages/lab/src/dialog/index.tsx
@@ -1,0 +1,6 @@
+export * from "./Dialog";
+export * from "./DialogTitle";
+export * from "./DialogActions";
+export * from "./DialogContent";
+export * from "./DialogContext";
+export * from "./DialogCloseButton";
