--- packages/lab/src/dialog/internal/DialogContext.ts
@@ -1,8 +1,0 @@
-import { createContext } from "react";
-import { ValidationStatus } from "@salt-ds/core";
-
-export const DialogContext = createContext<{
-  status?: ValidationStatus;
-  dialogId?: string;
-  setContentElement?: (node: HTMLDivElement) => void;
-}>({});
