+++ packages/lab/src/index.ts
@@ -29,7 +29,6 @@ export * from "./content-status";
 export * from "./deck-item";
 export * from "./deck-layout";
 export * from "./dialog";
-export * from "./dialog-next";
 export * from "./drawer";
 export * from "./dropdown";
 export * from "./editable-label";
