--- packages/lab/stories/dialog-next/dialog-next.stories.css
@@ -1,25 +1,0 @@
-nav {
-  position: relative;
-}
-
-ul {
-  list-style: none;
-  padding-left: 0;
-  margin: 0;
-}
-
-.verticalDialogNav {
-  width: 150px;
-}
-
-.fakeDialogWindow {
-  border: 1px solid black;
-  width: calc(20 * var(--salt-size-base));
-  padding-bottom: var(--salt-spacing-300);
-}
-
-.fakeDialogWindowHeader {
-  border-bottom: 1px solid black;
-  padding-block: var(--salt-spacing-200);
-  margin-bottom: var(--salt-spacing-300);
-}
