--- packages/lab/stories/dialog-next/dialog-next.stories.tsx
@@ -1,337 +1,0 @@
-import { PropsWithChildren, useState } from "react";
-import {
-  Button,
-  FlexLayout,
-  FlowLayout,
-  SplitLayout,
-  StackLayout,
-} from "@salt-ds/core";
-import {
-  DialogNext,
-  DialogNextTitle,
-  DialogNextActions,
-  DialogNextContent,
-  NavItem,
-  DialogNextCloseButton,
-} from "@salt-ds/lab";
-import { ComponentStory, ComponentMeta } from "@storybook/react";
-import "./dialog-next.stories.css";
-
-export default {
-  title: "Lab/Dialog Next",
-  component: DialogNext,
-  args: {
-    title: "Congratulations! You have created a Dialog.",
-    content: "This is the content of the dialog.",
-  },
-} as ComponentMeta<typeof DialogNext>;
-
-const DialogTemplate: ComponentStory<typeof DialogNext> = ({
-  title,
-  // @ts-ignore
-  accent,
-  // @ts-ignore
-  content,
-  open: openProp = true,
-  ...args
-}) => {
-  const [open, setOpen] = useState(openProp);
-
-  const handleRequestOpen = () => {
-    setOpen(true);
-  };
-
-  const onOpenChange = (value: boolean) => {
-    setOpen(value);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <>
-      <Button data-testid="dialog-button" onClick={handleRequestOpen}>
-        Click to open dialog
-      </Button>
-      <DialogNext {...args} open={open} onOpenChange={onOpenChange}>
-        <DialogNextTitle accent={accent as boolean}>{title}</DialogNextTitle>
-        <DialogNextContent>{content}</DialogNextContent>
-        <DialogNextActions>
-          <SplitLayout
-            style={{ flex: 1 }}
-            startItem={
-              <Button variant="secondary" onClick={handleClose}>
-                Cancel
-              </Button>
-            }
-            endItem={
-              <FlowLayout gap={1}>
-                <Button onClick={handleClose}>Previous</Button>
-                <Button variant="cta" onClick={handleClose}>
-                  Next
-                </Button>
-              </FlowLayout>
-            }
-          />
-        </DialogNextActions>
-        <DialogNextCloseButton onClick={handleClose} />
-      </DialogNext>
-    </>
-  );
-};
-
-export const Default = DialogTemplate.bind({});
-
-export const Accent = DialogTemplate.bind({});
-Accent.args = {
-  // @ts-ignore
-  accent: true,
-};
-
-export const LongContent = DialogTemplate.bind({});
-
-LongContent.args = {
-  title: "Congratulations! You have created a Dialog.",
-  // @ts-ignore
-  content: (
-    <StackLayout>
-      <div>
-        Lorem Ipsum is simply dummy text of the printing and typesetting
-        industry. Lorem Ipsum has been the industry's standard dummy text ever
-        since the 1500s, when an unknown printer took a galley of type and
-        scrambled it to make a type specimen book.
-      </div>
-      <div>
-        It has survived not only five centuries, but also the leap into
-        electronic typesetting, remaining essentially unchanged. It was
-        popularised in the 1960s with the release of Letraset sheets containing
-        Lorem Ipsum passages, and more recently with desktop publishing software
-        like Aldus PageMaker including versions of Lorem Ipsum.
-      </div>
-      <div>
-        It is a long established fact that a reader will be distracted by the
-        readable content of a page when looking at its layout. The point of
-        using Lorem Ipsum is that it has a more-or-less normal distribution of
-        letters, as opposed to using 'Content here, content here', making it
-        look like readable English.
-      </div>
-      <div>
-        Many desktop publishing packages and web page editors now use Lorem
-        Ipsum as their default model text, and a search for 'lorem ipsum' will
-        uncover many web sites still in their infancy. Various versions have
-        evolved over the years, sometimes by accident, sometimes on purpose
-        (injected humour and the like).
-      </div>
-      <div>
-        Contrary to popular belief, Lorem Ipsum is not simply random text. It
-        has roots in a piece of classical Latin literature from 45 BC, making it
-        over 2000 years old. Richard McClintock, a Latin professor at
-        Hampden-Sydney College in Virginia, looked up one of the more obscure
-        Latin words, consectetur, from a Lorem Ipsum passage, and going through
-        the cites of the word in classical literature, discovered the
-        undoubtable source.
-      </div>
-      <div>
-        Lorem Ipsum comes from sections 1.10.32 and 1.10.33 of "de Finibus
-        Bonorum et Malorum" (The Extremes of Good and Evil) by Cicero, written
-        in 45 BC. This book is a treatise on the theory of ethics, very popular
-        during the Renaissance. The first line of Lorem Ipsum, "Lorem ipsum
-        dolor sit amet..", comes from a line in section 1.10.32.
-      </div>
-    </StackLayout>
-  ),
-};
-
-const AlertDialogTemplate: ComponentStory<typeof DialogNext> = ({
-  open: openProp = true,
-  status,
-  ...args
-}) => {
-  const [open, setOpen] = useState(openProp);
-
-  const handleRequestOpen = () => {
-    setOpen(true);
-  };
-
-  const onOpenChange = (value: boolean) => {
-    setOpen(value);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <>
-      <Button data-testid="dialog-button" onClick={handleRequestOpen}>
-        Click to open dialog
-      </Button>
-      <DialogNext
-        {...args}
-        role="alertdialog"
-        status={status}
-        open={open}
-        onOpenChange={onOpenChange}
-        // focus the ok instead of the cancel button
-        initialFocus={1}
-      >
-        <DialogNextTitle>
-          {status![0].toUpperCase()}
-          {status!.slice(1)}
-        </DialogNextTitle>
-        <DialogNextContent>
-          This is a description for an alert dialog.
-        </DialogNextContent>
-        <DialogNextActions>
-          <Button variant="secondary" onClick={handleClose}>
-            Cancel
-          </Button>
-          <Button variant="primary" onClick={handleClose}>
-            Ok
-          </Button>
-        </DialogNextActions>
-      </DialogNext>
-    </>
-  );
-};
-
-export const InfoStatus = AlertDialogTemplate.bind({});
-InfoStatus.args = {
-  status: "info",
-};
-
-export const SuccessStatus = AlertDialogTemplate.bind({});
-SuccessStatus.args = {
-  status: "success",
-};
-
-export const WarningStatus = AlertDialogTemplate.bind({});
-WarningStatus.args = {
-  status: "warning",
-};
-
-export const ErrorStatus = AlertDialogTemplate.bind({});
-ErrorStatus.args = {
-  status: "error",
-};
-
-const itemsWithSubNav = [
-  {
-    name: "Nav Item 1",
-  },
-  {
-    name: "Nav Item 2",
-    subNav: ["Sub Nav Item 1", "Sub Nav Item 2", "Sub Nav Item 3"],
-  },
-  {
-    name: "Nav Item 3",
-    subNav: ["Sub Nav Item 1", "Sub Nav Item 2", "Sub Nav Item 3"],
-  },
-];
-
-const items = itemsWithSubNav.map((item) => item.name);
-
-export const PreferencesDialog: ComponentStory<typeof DialogNext> = (args) => {
-  const [open, setOpen] = useState(false);
-  const [active, setActive] = useState(items[0]);
-
-  const handleRequestOpen = () => {
-    setOpen(true);
-  };
-
-  const onOpenChange = (value: boolean) => {
-    setOpen(value);
-  };
-
-  const handleClose = () => {
-    setOpen(false);
-  };
-
-  return (
-    <>
-      <Button data-testid="dialog-button" onClick={handleRequestOpen}>
-        Click to open dialog
-      </Button>
-      <DialogNext
-        {...args}
-        open={open}
-        onOpenChange={onOpenChange}
-        style={{
-          width: "calc(20 * var(--salt-size-base))",
-        }}
-      >
-        <DialogNextTitle>Preferences</DialogNextTitle>
-        <FlexLayout gap={0}>
-          <nav>
-            <ul className="verticalDialogNav">
-              {items.map((item) => (
-                <li key={item}>
-                  <NavItem
-                    active={active === item}
-                    href="#"
-                    orientation="vertical"
-                    onClick={(event) => {
-                      // Prevent default to avoid navigation
-                      event.preventDefault();
-                      setActive(item);
-                    }}
-                  >
-                    {item}
-                  </NavItem>
-                </li>
-              ))}
-            </ul>
-          </nav>
-          <DialogNextContent>Hello</DialogNextContent>
-        </FlexLayout>
-        <DialogNextActions>
-          <Button onClick={handleClose}>Save</Button>
-        </DialogNextActions>
-        <DialogNextCloseButton onClick={handleClose} />
-      </DialogNext>
-    </>
-  );
-};
-
-function FakeWindow({ children }: PropsWithChildren) {
-  return (
-    <div className="fakeDialogWindow">
-      <div className="fakeDialogWindowHeader"></div>
-      {children}
-    </div>
-  );
-}
-
-export const DesktopDialog = () => {
-  return (
-    <StackLayout>
-      <FakeWindow>
-        <DialogNextTitle>Window Dialog</DialogNextTitle>
-        <DialogNextContent>Hello world!</DialogNextContent>
-        <DialogNextActions>
-          <Button variant="secondary">Cancel</Button>
-          <Button>Save</Button>
-        </DialogNextActions>
-      </FakeWindow>
-
-      <FakeWindow>
-        <DialogNextTitle accent>Window Dialog</DialogNextTitle>
-        <DialogNextContent>Accent world!</DialogNextContent>
-        <DialogNextActions>
-          <Button variant="secondary">Cancel</Button>
-          <Button>Save</Button>
-        </DialogNextActions>
-      </FakeWindow>
-
-      <FakeWindow>
-        <DialogNextTitle status="warning">Warning Dialog</DialogNextTitle>
-        <DialogNextContent>Potential issues abound!</DialogNextContent>
-        <DialogNextActions>
-          <Button variant="secondary">Cancel</Button>
-          <Button>Ok</Button>
-        </DialogNextActions>
-      </FakeWindow>
-    </StackLayout>
-  );
-};
