--- packages/lab/stories/dialog/dialog.qa.stories.css
@@ -1,3 +1,0 @@
-.examples-container .saltScrim {
-  position: unset;
-}
